# Optimize varint reading for Intel using intrinsics