BenchmarkDotNet v0.14.0, Windows 11 (10.0.26100.2314) Unknown processor
.NET SDK 9.0.100
  [Host]     : .NET 9.0.0 (9.0.24.52809), X64 RyuJIT AVX2
  DefaultJob : .NET 9.0.0 (9.0.24.52809), X64 RyuJIT AVX2

| Method   | Value | Mean     | Error     | StdDev    | Ratio | RatioSD | Rank |
|--------- |------ |---------:|----------:|----------:|------:|--------:|-----:|
| Previous | 0     | 3.881 ns | 0.0541 ns | 0.0506 ns |  1.00 |    0.02 |    2 |
| New      | 0     | 3.626 ns | 0.0312 ns | 0.0291 ns |  0.93 |    0.01 |    1 |
|          |       |          |           |           |       |         |      |
| Previous | 256   | 4.161 ns | 0.0096 ns | 0.0085 ns |  1.00 |    0.00 |    2 |
| New      | 256   | 3.596 ns | 0.0076 ns | 0.0067 ns |  0.86 |    0.00 |    1 |
|          |       |          |           |           |       |         |      |
| Previous | 65536 | 4.760 ns | 0.0103 ns | 0.0096 ns |  1.00 |    0.00 |    2 |
| New      | 65536 | 3.417 ns | 0.0087 ns | 0.0081 ns |  0.72 |    0.00 |    1 |